Module snowbridge_beacon_primitives::bls

Structs

AggregatePublicKey

Allows for the adding/combining of multiple BLS PublicKeys.

AggregateSignature

Allows for the adding/combining of multiple BLS Signatures.

PublicKeyPrepared

A BLS public key.

SignaturePrepared

Enums

BlsError

Functions

fast_aggregate_verify

fast_aggregate_verify optimized with aggregate key subtracting absent ones.

fast_aggregate_verify_pre_aggregated

fast_aggregate_verify_pre_aggregated which is the most expensive call in beacon light client.

prepare_aggregate_pubkey

Prepare for G1 AggregatePublicKey.

prepare_aggregate_pubkey_from_absent

Prepare for G1 AggregatePublicKey.

prepare_aggregate_signature

Prepare for G2 AggregateSignature, normally more expensive than G1 operation.

prepare_g1_pubkeys

Prepare for G1 public keys.

prepare_milagro_pubkey

Decompress one public key into a point in G1.