Module snowbridge_beacon_primitives::types

Re-exports

pub use crate::bits::decompress_sync_committee_bits;

Structs

BeaconHeader

Beacon block header as it is stored in the runtime storage. The block root is the Merkleization of a BeaconHeader.

CompactBeaconState

CompactExecutionHeader

ExecutionHeaderState

ExecutionPayloadHeader

ExecutionPayloadHeader https://github.com/ethereum/annotated-spec/blob/master/capella/beacon-chain.md#executionpayloadheader

FinalizedHeaderState

Fork

ForkData

ForkVersions

PublicKey

Signature

SigningData

SyncAggregate

SyncCommittee

Sync committee as it is stored in the runtime storage.

SyncCommitteePrepared

Prepared G1 public key of sync committee as it is stored in the runtime storage.

Enums

Mode

Operating modes for beacon client

Type Aliases

ForkVersion

ValidatorIndex